LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/clipboard-apis/async-navigator-clipboard-basics.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://localhost:9443/clipboard-apis/async-navigator-clipboard-basics.https.html was not allowed to display insecure content from http://localhost:8001/clipboard-apis/resources/greenbox.png.

CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8001/clipboard-apis/resources/greenbox.png due to access control checks.

FAIL navigator.clipboard exists assert_not_equals: got disallowed value undefined
FAIL navigator.clipboard.write([text/plain ClipboardItem]) succeeds promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: ClipboardItem"
FAIL navigator.clipboard.write() fails (expect [ClipboardItem]) promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.write')"
FAIL navigator.clipboard.write(null) fails (expect [ClipboardItem]) promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.write')"
FAIL navigator.clipboard.write(DOMString) fails (expect [ClipboardItem]) promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.write')"
FAIL navigator.clipboard.write(Blob) fails (expect [ClipboardItem]) promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.write')"
FAIL navigator.clipboard.writeText(DOMString) succeeds promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.writeText')"
FAIL navigator.clipboard.writeText() fails (expect DOMString) promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.writeText')"
FAIL navigator.clipboard.write({string : image/png Blob}) succeeds promise_test: Unhandled rejection with value: object "TypeError: Not allowed to request resource"
FAIL navigator.clipboard.read() succeeds promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.read')"
FAIL navigator.clipboard.readText() succeeds promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'navigator.clipboard.readText')"
